requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
altair==5.5.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
bcrypt==3.2.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
fastapi==0.120.0
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.11.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langdetect==1.0.9
MarkupSafe==3.0.3
multidict==6.7.0
narwhals==2.9.0
numpy==2.3.4
openai==0.28.0
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pillow==11.3.0
propcache==0.4.1
protobuf==6.33.0
psycopg2-binary==2.9.11
pyarrow==22.0.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pydeck==0.9.1
pymongo==4.15.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.28.0
rsa==4.9.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.48.0
streamlit==1.50.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
watchdog==6.0.0
yarl==1.22.0
